+++ src/main/kotlin/com/github/erotourtes/harpoon/utils/menu/QuickMenu.kt
@@ -223,7 +223,6 @@ class QuickMenu(private val project: Project, private val harpoonService: Harpoo
             if (!isHarpoonerPrevFocused || isRefocusOnMenu) return
 
             harpoonService.syncWithMenu()
-            foldManager.collapseAllFolds()
             closeMenuInEditor()
             isHarpoonerPrevFocused = false
         }